Move su-gitlogcheck command into a separate file mx_gitlogcheck.py
